Revert "-try to fixed notification color"
This reverts commit b3135bb9cc461d206c046998ff82fe9c8214defc.

+++ flutter-hms-push/android/src/main/java/com/huawei/hms/flutter/push/localnotification/HmsLocalNotificationController.java
@@ -326,7 +326,7 @@ public class HmsLocalNotificationController {
 
                 String color = BundleUtils.get(bundle, NotificationConstants.COLOR);
                 if (color != null) {
-                    notification.setColor(Color.parseColor(color).intValue());
+                    notification.setColor(Color.parseColor(color));
                 }
             }
